+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 
 - Privatized renderer backend
 - Small optimizations on renderer sorting and copying sprites
+- Expanded Texture loading to instead use `zune_png` and `zune_jpeg`
 
 ## v0.1.0 (24/07/2026)